pandaproxy/rest_proxy: Move into a subfolder
