+++ src/v/pandaproxy/CMakeLists.txt
@@ -1,25 +1,9 @@
-set (swagger_deps
-  v1
-)
-
-foreach(swagger_dep ${swagger_deps})
-  seastar_generate_swagger(
-    TARGET ${swagger_dep}_swagger
-    VAR ${swagger_dep}_file
-    IN_FILE ${CMAKE_CURRENT_SOURCE_DIR}/api/api-doc/${swagger_dep}.json
-    OUT_FILE ${CMAKE_CURRENT_BINARY_DIR}/api/api-doc/${swagger_dep}.json.h
-  )
-endforeach()
-
 v_cc_library(
-  NAME pandaproxy
+  NAME pandaproxy_common
   SRCS
-    configuration.cc
     error.cc
-    handlers.cc
     logger.cc
     probe.cc
-    proxy.cc
     server.cc
   DEPS
     v::pandaproxy_parsing
@@ -31,10 +15,7 @@ v_cc_library(
     v::utils
   )
 
-foreach(swagger_dep ${swagger_deps})
-  add_dependencies(v_pandaproxy ${swagger_dep}_swagger)
-endforeach()
 
+add_subdirectory(rest)
 add_subdirectory(json)
 add_subdirectory(parsing)
-add_subdirectory(test)
