+++ src/v/pandaproxy/api/api-doc/rest.json
