+++ src/v/pandaproxy/json/requests/test/CMakeLists.txt
@@ -5,6 +5,6 @@ rp_test(
     fetch.cc
     produce.cc
   DEFINITIONS BOOST_TEST_DYN_LINK
-  LIBRARIES v::seastar_testing_main v::pandaproxy
+  LIBRARIES v::seastar_testing_main v::pandaproxy_rest v::utils
   LABELS pandaproxy
 )
