+++ src/v/pandaproxy/json/test/CMakeLists.txt
@@ -4,6 +4,6 @@ rp_test(
   SOURCES
     iobuf.cc
   DEFINITIONS BOOST_TEST_DYN_LINK
-  LIBRARIES v::seastar_testing_main v::pandaproxy
+  LIBRARIES v::seastar_testing_main v::pandaproxy_json v::utils
   LABELS pandaproxy
 )
