+++ src/v/pandaproxy/rest/configuration.cc
