+++ src/v/pandaproxy/rest/fwd.h
