+++ src/v/pandaproxy/rest/handlers.cc
@@ -18,7 +18,6 @@
 #include "kafka/protocol/schemata/offset_commit_request.h"
 #include "kafka/types.h"
 #include "model/fundamental.h"
-#include "pandaproxy/configuration.h"
 #include "pandaproxy/json/exceptions.h"
 #include "pandaproxy/json/requests/create_consumer.h"
 #include "pandaproxy/json/requests/fetch.h"
@@ -31,6 +30,7 @@
 #include "pandaproxy/json/rjson_util.h"
 #include "pandaproxy/parsing/httpd.h"
 #include "pandaproxy/reply.h"
+#include "pandaproxy/rest/configuration.h"
 #include "raft/types.h"
 #include "ssx/future-util.h"
 #include "ssx/sformat.h"
