+++ src/v/pandaproxy/server.cc
@@ -11,11 +11,11 @@
 
 #include "cluster/cluster_utils.h"
 #include "model/metadata.h"
-#include "pandaproxy/configuration.h"
 #include "pandaproxy/json/types.h"
 #include "pandaproxy/logger.h"
 #include "pandaproxy/probe.h"
 #include "pandaproxy/reply.h"
+#include "pandaproxy/rest/configuration.h"
 
 #include <seastar/core/coroutine.hh>
 #include <seastar/core/std-coroutine.hh>
