+++ src/v/redpanda/CMakeLists.txt
@@ -51,7 +51,7 @@ v_cc_library(
     v::syschecks
     v::kafka
     v::coproc
-    v::pandaproxy
+    v::pandaproxy_rest
     v::archival
   )
