+++ src/v/redpanda/application.cc
@@ -33,8 +33,8 @@
 #include "kafka/server/queue_depth_monitor.h"
 #include "kafka/server/quota_manager.h"
 #include "model/metadata.h"
-#include "pandaproxy/configuration.h"
-#include "pandaproxy/proxy.h"
+#include "pandaproxy/rest/configuration.h"
+#include "pandaproxy/rest/proxy.h"
 #include "platform/stop_signal.h"
 #include "raft/service.h"
 #include "redpanda/admin_server.h"
