+++ src/v/redpanda/application.h
@@ -19,8 +19,8 @@
 #include "coproc/pacemaker.h"
 #include "kafka/server/fwd.h"
 #include "kafka/server/rm_group_frontend.h"
-#include "pandaproxy/configuration.h"
-#include "pandaproxy/fwd.h"
+#include "pandaproxy/rest/configuration.h"
+#include "pandaproxy/rest/fwd.h"
 #include "raft/group_manager.h"
 #include "redpanda/admin_server.h"
 #include "resource_mgmt/cpu_scheduling.h"
